requirements.txt
streamlit
transformers
torch
pdfminer.six
nltk
scikit-learn
sentencepiece
